+++ data/event/event.go
@@ -8,8 +8,9 @@ const (
 // Event is the data type to represent arbitrary, one-off messages for key
 // activities on a system.
 type Event struct {
-	Summary  string `json:"summary"`
-	Category string `json:"category,omitempty"`
+	Summary    string                 `json:"summary"`
+	Category   string                 `json:"category,omitempty"`
+	Attributes map[string]interface{} `json:"attributes,omitempty"`
 }
 
 // New creates a new event.
@@ -23,4 +24,11 @@ func New(summary, category string) *Event {
 // NewNotification creates a new notification event.
 func NewNotification(summary string) *Event {
 	return New(summary, NotificationEventCategory)
+}
+
+// NewWithAttributes creates a new event with the given attributes
+func NewWithAttributes(summary, category string, attributes map[string]interface{}) *Event {
+	e := New(summary, category)
+	e.Attributes = attributes
+	return e
 }