+++ data/event/event_test.go
@@ -17,3 +17,15 @@ func TestNewNotification(t *testing.T) {
 	n := NewNotification("summary")
 	assert.Equal(t, n.Summary, "summary")
 }
+
+func TestNewAttributes(t *testing.T) {
+	e := NewWithAttributes(
+		"summary",
+		"category",
+		map[string]interface{}{"attrKey": "attrVal"},
+	)
+
+	assert.Equal(t, e.Summary, "summary")
+	assert.Equal(t, e.Category, "category")
+	assert.Equal(t, e.Attributes["attrKey"], "attrVal")
+}